+++ ChangeLog.md
@@ -6,6 +6,7 @@
 
 - Added `JumpToSubscriptions`, `JumpToArticles` and `JumpToArticle` commands
   for quickly jumping between the 3 main panels.
+  ([#197](https://github.com/davep/oldnews/pull/197))
 
 ## v1.3.0